dom/tests/mochitest/sessionstorage/test_sessionStorageBase.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>sessionStorage basic test</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

var INDEX_SIZE_ERR = 1;

function checkException(func, exc)
{
  var exceptionThrew = false;
  try {
    func();
  }
  catch (ex) {
    exceptionThrew = true;
    is(ex.code, exc, "Expected "+exc+" exception");
  }
  ok(exceptionThrew, "Exception "+exc+" threw");
}

function startTest()
{
  sessionStorage.clear();
  
  // Initially check the sessionStorage is empty
  is(sessionStorage.length, 0, "The storage is empty [1]");
  checkException(function() {sessionStorage.key(0);}, INDEX_SIZE_ERR);
  checkException(function() {sessionStorage.key(-1);}, INDEX_SIZE_ERR);
  checkException(function() {sessionStorage.key(1);}, INDEX_SIZE_ERR);
  is(sessionStorage.getItem("nonexisting"), null, "Nonexisting item is null (getItem())");
  is(sessionStorage["nonexisting"], null, "Nonexisting item is null (array access)");
  is(sessionStorage.nonexisting, null, "Nonexisting item is null (property access)");
  sessionStorage.removeItem("nonexisting"); // Just check there is no exception

  is(typeof sessionStorage.getItem("nonexisting"), "object", "getItem('nonexisting') is object");
  is(typeof sessionStorage["nonexisting"], "object", "['nonexisting'] is object");
  is(typeof sessionStorage.nonexisting, "object", "nonexisting is object");
  is(typeof sessionStorage.getItem("nonexisting2"), "object", "getItem('nonexisting2') is object");
  is(typeof sessionStorage["nonexisting2"], "object", "['nonexisting2'] is object");
  is(typeof sessionStorage.nonexisting2, "object", "nonexisting2 is object");

  // add an empty-value key
  sessionStorage.setItem("empty", "");
  is(sessionStorage.getItem("empty"), "", "Empty value (getItem())");
  is(sessionStorage["empty"], "", "Empty value (array access)");
  is(sessionStorage.empty, "", "Empty value (property access)");
  is(typeof sessionStorage.getItem("empty"), "string", "getItem('empty') is string");
  is(typeof sessionStorage["empty"], "string", "['empty'] is string");
  is(typeof sessionStorage.empty, "string", "empty is string");
  sessionStorage.removeItem("empty");
  is(sessionStorage.length, 0, "The storage has no keys");
  is(sessionStorage.getItem("empty"), null, "empty item is null (getItem())");
  is(sessionStorage["empty"], null, "empty item is null (array access)");
  is(sessionStorage.empty, null, "empty item is null (property access)");
  is(typeof sessionStorage.getItem("empty"), "object", "getItem('empty') is object");
  is(typeof sessionStorage["empty"], "object", "['empty'] is object");
  is(typeof sessionStorage.empty, "object", "empty is object");

  // add one key, check it is there
  sessionStorage.setItem("key1", "value1");
  is(sessionStorage.length, 1, "The storage has one key-value pair");
  is(sessionStorage.key(0), "key1");
  checkException(function() {sessionStorage.key(-1);}, INDEX_SIZE_ERR);
  checkException(function() {sessionStorage.key(1);}, INDEX_SIZE_ERR);

  // check all access method give the correct result
  // and are of the correct type
  is(sessionStorage.getItem("key1"), "value1", "getItem('key1') == value1");
  is(sessionStorage["key1"], "value1", "['key1'] == value1");
  is(sessionStorage.key1, "value1", "key1 == value1");

  is(typeof sessionStorage.getItem("key1"), "string", "getItem('key1') is string");
  is(typeof sessionStorage["key1"], "string", "['key1'] is string");
  is(typeof sessionStorage.key1, "string", "key1 is string");

  // remove the previously added key and check the storage is empty
  sessionStorage.removeItem("key1");
  is(sessionStorage.length, 0, "The storage is empty [2]");
  checkException(function() {sessionStorage.key(0);}, INDEX_SIZE_ERR);
  is(sessionStorage.getItem("key1"), null, "\'key1\' removed");

  is(typeof sessionStorage.getItem("key1"), "object", "getItem('key1') is object");
  is(typeof sessionStorage["key1"], "object", "['key1'] is object");
  is(typeof sessionStorage.key1, "object", "key1 is object");

  // add one key, check it is there
  sessionStorage.setItem("key1", "value1");
  is(sessionStorage.length, 1, "The storage has one key-value pair");
  is(sessionStorage.key(0), "key1");
  is(sessionStorage.getItem("key1"), "value1");

  // add a second key
  sessionStorage.setItem("key2", "value2");
  is(sessionStorage.length, 2, "The storage has two key-value pairs");
  is(sessionStorage.getItem("key1"), "value1");
  is(sessionStorage.getItem("key2"), "value2");
  var firstKey = sessionStorage.key(0);
  var secondKey = sessionStorage.key(1);
  ok((firstKey == 'key1' && secondKey == 'key2') ||
     (firstKey == 'key2' && secondKey == 'key1'),
     'Both keys should be present.');

  // change the second key
  sessionStorage.setItem("key2", "value2-2");
  is(sessionStorage.length, 2, "The storage has two key-value pairs");
  is(sessionStorage.key(0), firstKey); // After key value changes the order must be preserved
  is(sessionStorage.key(1), secondKey);
  checkException(function() {sessionStorage.key(-1);}, INDEX_SIZE_ERR);
  checkException(function() {sessionStorage.key(2);}, INDEX_SIZE_ERR);
  is(sessionStorage.getItem("key1"), "value1");
  is(sessionStorage.getItem("key2"), "value2-2");

  // change the first key
  sessionStorage.setItem("key1", "value1-2");
  is(sessionStorage.length, 2, "The storage has two key-value pairs");
  is(sessionStorage.key(0), firstKey); // After key value changes the order must be preserved
  is(sessionStorage.key(1), secondKey);
  checkException(function() {sessionStorage.key(-1);}, INDEX_SIZE_ERR);
  checkException(function() {sessionStorage.key(2);}, INDEX_SIZE_ERR);
  is(sessionStorage.getItem("key1"), "value1-2");
  is(sessionStorage.getItem("key2"), "value2-2");

  // remove the second key
  sessionStorage.removeItem("key2");
  is(sessionStorage.length, 1, "The storage has one key-value pair");
  is(sessionStorage.key(0), "key1");
  checkException(function() {sessionStorage.key(-1);}, INDEX_SIZE_ERR);
  checkException(function() {sessionStorage.key(1);}, INDEX_SIZE_ERR);
  is(sessionStorage.getItem("key1"), "value1-2");

  // Clear the storage
  sessionStorage.clear();
  is(sessionStorage.length, 0, "The storage is empty [3]");
  checkException(function() {sessionStorage.key(0);}, INDEX_SIZE_ERR); // this is unspecified!
  checkException(function() {sessionStorage.key(-1);}, INDEX_SIZE_ERR);
  checkException(function() {sessionStorage.key(1);}, INDEX_SIZE_ERR);
  is(sessionStorage.getItem("nonexisting"), null, "Nonexisting item is null");
  is(sessionStorage.getItem("key1"), null, "key1 removed");
  is(sessionStorage.getItem("key2"), null, "key2 removed");
  sessionStorage.removeItem("nonexisting"); // Just check there is no exception
  sessionStorage.removeItem("key1"); // Just check there is no exception
  sessionStorage.removeItem("key2"); // Just check there is no exception

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body onload="startTest();">

</body>
</html>
